EMLeditor: View and Edit EML Metadata

This package will be of most use to the U.S. National Park Service data scientists and managers seeking to generate EML-formatted metadata for datapackages. EML-formatted .xml files are typically constructed using EDI's EMLassemblyline package and then imported as an R-object using the EML package. EMLeditor allows the user to view the contents of the R object and add/edit aspects of metadata crucial for publication in the U.S. National Park Service DataStore repository. For instance, a user can view and edit a DOI, a link to a DRR, Park Unit connections, information about Confidential Unclassified Information (CUI), and more. EMLeditor allows the user to write a mockup of a README.txt to preview what the README automatically generated by DataStore upon upload will look like.

See also

Author

Maintainer: Robert Baker robert_baker@nps.gov (ORCID)

Authors:

Judd Patterson Judd_Patterson@nps.gov (ORCID)

Other contributors:

Issac Quevedo (ORCID) [contributor]

Amy Sherman (ORCID) [contributor]