Adds CUI to metadata

#' set_cui adds CUI dissemination codes to EML metadata

Usage

set_cui(
  eml_object,
  cui_code = c("PUBLIC", "NOCON", "DL ONLY", "FEDCON", "FED ONLY"),
  force = FALSE,
  NPS = TRUE
)

Arguments

eml_object

is an EML-formatted R object, either generated in R or imported (typically from an EML-formatted .xml file) using EML::read_eml(, from="xml").

cui_code

a string consisting of one of 5 potential CUI codes. Pay attention to the spaces: FED ONLY - Contains CUI. Only federal employees should have access (similar to "internal only" in DataStore) FEDCON - Contains CUI. Only federal employees and federal contractors should have access (also very much like current "internal only" setting in DataStore) DL ONLY - Contains CUI. Should only be available to a names list of individuals (where and how to list those individuals TBD) NOCON - Contains CUI. Federal, state, local, or tribal employees may have access, but contractors cannot. PUBLIC - Does NOT contain CUI.

force

logical. Defaults to false. If set to FALSE, a more interactive version of the function requesting user input and feedback. Setting force = TRUE facilitates scripting.

NPS

Logical. Defaults to TRUE. Most NPS users should leave this as the default. Only under specific circumstances should it be set to FALSE: if you are not publishing with NPS, if you need to set the publisher location to some place other than the Fort Collins Office (e.g. you are NOT working on a data package) or your product is "for" the NPS but not "by" the NPS and you need to specify a different agency, set NPS = FALSE. When NPS=TRUE, the function will over-write existing publisher info and inject NPS as the publisher along the the Central Office in Fort Collins as the location. Additionally, it sets the "for or by NPS" field to TRUE and specifies the originating agency as NPS.

Value

an EML-formatted R object

Details

set_cui adds a CUI code to the tag CUI under additionalMetadata/metadata.

Examples

if (FALSE) { # \dontrun{
set_cui(eml_object, "PUBFUL")
} # }